+++ src/helpers/types.ts
@@ -58,7 +58,11 @@ export function convertToType(Target: any, data?: object): object | undefined {
   if (data == null) {
     return;
   }
-  // skip simple types
+  // skip converting scalars (object scalar mostly)
+  if (Target instanceof GraphQLScalarType) {
+    return data;
+  }
+  // skip converting simple types
   if (simpleTypes.includes(data.constructor)) {
     return data;
   }